backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
textstat>=0.7.0
openai>=1.0.0
groq>=0.4.0
requests>=2.28.0
python-dotenv>=1.0.0
